Add files via upload
made pid controller modular, and for left and right turns, all we need to do is call a target velocity for left and right wheels opposite each other for a certain amount of time. forward works, left and right not fully tested, kinda works? On a note, m1 motor 30 and m2 motor 34 PWM lets it move forward straight, i don't how to get the arduino to retain this info. Also, just find+replace all the nano34 code and pins with heltech to run on new board.