Add files via upload
turned off left turn testing. mostly fine tuned the pid controller so that m2 doesn't kick at max pwm on each loop start up. This caused it to over shoot the position comparison between curr and target. m1 oscillates in it's velocity, need to dampen d and jack up i. They are not moving at the same speed, but at least reasonable levels and not maxed out. Additionally for the turning code, it might not be necessary to have it be controlled by the pid_controller, and instead just default max spd to one motor, and 0 to the other, so all we need to do is calculate pivot angle. It's mostly in anticipation for curved turn line following, otherwise straight edge turns is easy.