Add files via upload
The lighting of my room isn't the same as the lab, so the movement is temp deleted. But the pid control is all there. Additionally, it's a PD controller instead of PID, because with an integrator component, there is some complexity that isn't necessary right now. Additionally, the response of the line sensor is super inaccurate sometimes reading absolutely nothing even though it's on the line, values are set to be 0 if that's the case, but it can result in choppy movement. Will implement the AR model on to kpid tuning later this week after I figure out how to get positional value from encoder + speed.